**structural project refactoring :: added gen0 default nn model
