+++ src/GammonX/GammonX.Mars.NN.Tests/Data/MockBoardContracts.cs
@@ -1,0 +1,49 @@
+﻿namespace GammonX.Mars.NN.Tests.Data
+{
+    public static class MockBoards
+    {
+        #region Test Plakoto Boards
+
+        public const string PlakotoBorneOffBoard1 = "{\"fields\":[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],\"bearOffCountWhite\":15,\"bearOffCountBlack\":15,\"pipCountWhite\":350,\"pipCountBlack\":350,\"pinnedFields\":[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]}";
+
+        public const string PlakotoBoard1 = "{\"fields\":[-13,0,0,0,-1,0,-1,0,0,0,0,0,0,0,0,0,0,1,0,1,0,0,0,13],\"bearOffCountWhite\":0,\"bearOffCountBlack\":0,\"pipCountWhite\":350,\"pipCountBlack\":350,\"pinnedFields\":[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]}";
+
+        public const string PlakotoBoard2 = "{\"fields\":[-13,0,0,0,-1,0,-1,0,0,0,0,0,0,0,0,0,1,0,1,0,0,0,0,13],\"bearOffCountWhite\":0,\"bearOffCountBlack\":0,\"pipCountWhite\":350,\"pipCountBlack\":348,\"pinnedFields\":[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]}";
+
+        public const string PlakotoBoard3 = "{\"fields\":[-13,0,0,-1,-1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,12],\"bearOffCountWhite\":0,\"bearOffCountBlack\":0,\"pipCountWhite\":350,\"pipCountBlack\":350,\"pinnedFields\":[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]}";
+
+        public const string PlakotoBoard4 = "{\"fields\":[-6,0,0,0,1,-2,2,0,0,0,-2,-1,0,0,0,0,0,-2,-1,3,0,0,0,8],\"bearOffCountWhite\":0,\"bearOffCountBlack\":0,\"pipCountWhite\":350,\"pipCountBlack\":350,\"pinnedFields\":[0,0,0,0,0,0,-1,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0]}";
+
+        public const string PinPlakotoBoard1 = "{\"fields\":[-10,1,1,0,-1,0,-1,0,0,0,0,0,0,0,0,0,0,1,1,1,0,-1,-1,9],\"bearOffCountWhite\":0,\"bearOffCountBlack\":0,\"pipCountWhite\":350,\"pipCountBlack\":350,\"pinnedFields\":[0,0,-1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0]}";
+
+        public const string PinPlakotoBoard2 = "{\"fields\":[1,-1,1,-6,-1,0,-1,0,0,0,0,0,0,0,0,0,0,1,-1,1,6,-1,1,-1],\"bearOffCountWhite\":0,\"bearOffCountBlack\":0,\"pipCountWhite\":350,\"pipCountBlack\":350,\"pinnedFields\":[-1,1,-1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,-1,1]}";
+
+        public const string PlakotoToBearOffBoard1 = "{\"fields\":[-13,0,0,0,-1,0,-1,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,13],\"bearOffCountWhite\":0,\"bearOffCountBlack\":0,\"pipCountWhite\":350,\"pipCountBlack\":350,\"pinnedFields\":[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]}";
+
+        public const string PlakotoRaceBoard1 = "{\"fields\":[0,0,0,0,0,0,0,0,0,0,0,0,0,15,-15,0,0,0,0,0,0,0,0,0],\"bearOffCountWhite\":0,\"bearOffCountBlack\":0,\"pipCountWhite\":350,\"pipCountBlack\":350,\"pinnedFields\":[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]}";
+
+        public const string PlakotoRacePinBoard1 = "{\"fields\":[0,0,0,0,0,0,0,0,0,0,0,0,0,14,-14,0,0,0,0,0,-1,0,0,0],\"bearOffCountWhite\":0,\"bearOffCountBlack\":0,\"pipCountWhite\":350,\"pipCountBlack\":350,\"pinnedFields\":[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0]}";
+
+        public const string PlakotoBlackWonBoard = "{\"fields\":[-15,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],\"bearOffCountWhite\":0,\"bearOffCountBlack\":15,\"pipCountWhite\":350,\"pipCountBlack\":0,\"pinnedFields\":[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]}";
+
+        #endregion Test Plakoto Boards
+
+        #region Test Fevga Boards
+
+        public const string FevgaBlackWonBoard = "{\"fields\":[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],\"bearOffCountWhite\":0,\"bearOffCountBlack\":15,\"pipCountWhite\":350,\"pipCountBlack\":350,\"homebarCountWhite\":15,\"homebarCountBlack\":0}";
+
+        public const string FevgaBoard1 = "{\"fields\":[-1,-2,0,-2,-2,-2,0,-1,0,0,0,0,2,2,2,1,0,3,3,2,0,0,0,0],\"bearOffCountWhite\":0,\"bearOffCountBlack\":0,\"pipCountWhite\":350,\"pipCountBlack\":350,\"homebarCountWhite\":5,\"homebarCountBlack\":0}";
+
+        public const string FevgaBoard2 = "{\"fields\":[-3,0,0,-2,-1,0,1,0,0,0,0,0,1,1,1,1,1,-1,0,0,0,0,0,1],\"bearOffCountWhite\":0,\"bearOffCountBlack\":0,\"pipCountWhite\":333,\"pipCountBlack\":321,\"pinnedFields\":null,\"homeBarCountWhite\":8,\"homeBarCountBlack\":8,\"doublingCubeValue\":0,\"doublingCubeOwner\":false}";
+
+        public const string FevgaBoard3 = "{\"fields\":[1,-1,1,0,2,-14,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],\"bearOffCountWhite\":0,\"bearOffCountBlack\":11,\"pipCountWhite\":350,\"pipCountBlack\":350,\"homebarCountWhite\":0,\"homebarCountBlack\":0}";
+
+        #endregion Test Fevga Boards
+
+        #region Default Boards
+
+        public const string DefaultBoard1 = "{\"fields\":[-1,-2,0,-2,-2,-2,0,-1,0,0,0,0,2,2,2,1,0,3,3,2,0,0,0,0],\"bearOffCountWhite\":0,\"bearOffCountBlack\":0,\"pipCountWhite\":350,\"pipCountBlack\":350,\"homebarCountWhite\":5,\"homebarCountBlack\":0}";
+
+        #endregion
+    }
+}
