+++ src/GammonX/GammonX.Mars.NN.Tests/Features/BlotCountFeatureTests.cs
@@ -2,16 +2,16 @@
 
 using GammonX.Mars.NN.Features;
 
-using GammonX.Mars.Server.Tests.Data;
+using GammonX.Mars.NN.Tests.Data;
 
 using GammonX.Models.Contracts;
 using GammonX.Models.Enums;
 
 using Newtonsoft.Json;
 
-namespace GammonX.Mars.Server.Tests.Features
+namespace GammonX.Mars.NN.Tests.Features
 {
-    public class BlotInStartRangeCountFeatureTests
+    public class BlotCountFeatureTests
     {
         [Fact]
         public void CanEvalPinPlakotoBoard1()
@@ -21,11 +21,11 @@ namespace GammonX.Mars.Server.Tests.Features
             Assert.NotNull(boardContract);
             var board = boardService.CreateBoard(boardContract);
 
-            var feature = new BlotInStartRangeCountFeature();
+            var feature = new BlotCountFeature();
             var whiteResult = feature.Eval(board, true);
-            Assert.Equal(1, whiteResult);
+            Assert.Equal(3, whiteResult);
             var blackResult = feature.Eval(board, false);
-            Assert.Equal(2, blackResult);
+            Assert.Equal(4, blackResult);
         }
 
         [Fact]
@@ -36,11 +36,26 @@ namespace GammonX.Mars.Server.Tests.Features
             Assert.NotNull(boardContract);
             var board = boardService.CreateBoard(boardContract);
 
-            var feature = new BlotInStartRangeCountFeature();
+            var feature = new BlotCountFeature();
             var whiteResult = feature.Eval(board, true);
-            Assert.Equal(1, whiteResult);
+            Assert.Equal(2, whiteResult);
             var blackResult = feature.Eval(board, false);
-            Assert.Equal(1, blackResult);
+            Assert.Equal(2, blackResult);
+        }
+
+        [Fact]
+        public void CanEvalFevgaBoard1()
+        {
+            var boardService = BoardServiceFactory.Create(GameModus.Fevga);
+            var boardContract = JsonConvert.DeserializeObject<BoardModelContract>(MockBoards.FevgaBoard1);
+            Assert.NotNull(boardContract);
+            var board = boardService.CreateBoard(boardContract);
+
+            var feature = new BlotCountFeature();
+            var result = feature.Eval(board, true);
+            Assert.Equal(2, result);
+            result = feature.Eval(board, false);
+            Assert.Equal(1, result);
         }
     }
 }
