+++ src/GammonX/GammonX.Mars.NN.Tests/Features/MotherDistanceFeatureTests.cs
@@ -2,45 +2,45 @@
 
 using GammonX.Mars.NN.Features;
 
-using GammonX.Mars.Server.Tests.Data;
+using GammonX.Mars.NN.Tests.Data;
 
 using GammonX.Models.Contracts;
 using GammonX.Models.Enums;
 
 using Newtonsoft.Json;
 
-namespace GammonX.Mars.Server.Tests.Features
+namespace GammonX.Mars.NN.Tests.Features
 {
-    public class AverageStackHeightFeatureTests
+    public class MotherDistanceFeatureTests
     {
         [Fact]
-        public void CanEvalPlakotoBoard1()
+        public void CanEvalPinPlakotoBoard1()
         {
             var boardService = BoardServiceFactory.Create(GameModus.Plakoto);
-            var boardContract = JsonConvert.DeserializeObject<BoardModelContract>(MockBoards.PlakotoBoard1);
+            var boardContract = JsonConvert.DeserializeObject<BoardModelContract>(MockBoards.PinPlakotoBoard1);
             Assert.NotNull(boardContract);
             var board = boardService.CreateBoard(boardContract);
 
-            var feature = new AverageStackHeightFeature();
+            var feature = new MotherDistanceFeature();
             var whiteResult = feature.Eval(board, true);
-            Assert.Equal(5, whiteResult);
+            Assert.Equal(18, whiteResult);
             var blackResult = feature.Eval(board, false);
-            Assert.Equal(5, blackResult);
+            Assert.Equal(18, blackResult);
         }
 
         [Fact]
-        public void CanEvalFevgaBoard1()
+        public void CanEvalPlakotoBorneOffBoard1()
         {
-            var boardService = BoardServiceFactory.Create(GameModus.Fevga);
-            var boardContract = JsonConvert.DeserializeObject<BoardModelContract>(MockBoards.FevgaBoard1);
+            var boardService = BoardServiceFactory.Create(GameModus.Plakoto);
+            var boardContract = JsonConvert.DeserializeObject<BoardModelContract>(MockBoards.PlakotoBorneOffBoard1);
             Assert.NotNull(boardContract);
             var board = boardService.CreateBoard(boardContract);
 
-            var feature = new AverageStackHeightFeature();
+            var feature = new MotherDistanceFeature();
             var whiteResult = feature.Eval(board, true);
-            Assert.Equal(1.6666666666666667, whiteResult);
+            Assert.Equal(0, whiteResult);
             var blackResult = feature.Eval(board, false);
-            Assert.Equal(2.142857142857142857142857142857, blackResult);
+            Assert.Equal(0, blackResult);
         }
 
         [Fact]
@@ -51,9 +51,9 @@ namespace GammonX.Mars.Server.Tests.Features
             Assert.NotNull(boardContract);
             var board = boardService.CreateBoard(boardContract);
 
-            var feature = new AverageStackHeightFeature();
+            var feature = new MotherDistanceFeature();
             var whiteResult = feature.Eval(board, true);
-            Assert.Equal(15, whiteResult);
+            Assert.Equal(18, whiteResult);
             var blackResult = feature.Eval(board, false);
             Assert.Equal(0, blackResult);
         }
