--- src/GammonX/GammonX.Mars.Server.Tests/Data/NeuralNets/Fevga/README.md
@@ -1,32 +1,0 @@
-### Trainings Data
-|    Metric    |    Value     |
-| ------------ | ------------ |
-| prior gen    | gen8         |
-| sample size  | 80_000       |
-| bootstrap    | 1            |
-| completed    | 80_000       |
-| discarded    | 0            |
-| samples      | 7_551_680    |
-| avg turns    | 96.3         |
-| avg pred var | 0.10780      |
-| duration     | 00:14:15:36  |
-
-### Training Model
-|    Metric    |    Value     |
-| ------------ | ------------ |
-| train data   | 9            |
-| train size   | 6_418_928    |
-| val size     | 1_132_752    |
-| train mean   | 0.5112       |
-| train min    | 0            |
-| train max    | 1            |
-| t-near 0.5   | 0.0%         |
-| val mean     | 0.5118       |
-| val min      | 0            |
-| val max      | 1            |
-| v-near 0.5   | 0.0%         |
-| e1 t-loss    | 0.43163      |
-| e1 v-loss    | 0.41740      |
-| e22 t-loss   | 0.40867      |
-| e22 v-loss   | 0.40762      |
-| val gap      | 0.00105      |
