--- src/GammonX/GammonX.Mars.Server.Tests/Data/NeuralNets/Plakoto/README.md
@@ -1,54 +1,0 @@
-### Trainings Data
-|    Metric    |    Value     |
-| ------------ | ------------ |
-| prior gen    | gen8         |
-| sample size  | 80_000       |
-| bootstrap    | 1            |
-| completed    | 80_000       |
-| discarded    | 0            |
-| samples      | 7_494_731    |
-| avg turns    | 98.0         |
-| avg pred var | 0.13560      |
-| duration     | 00:17:25:40  |
-
-### Training Model
-|    Metric    |    Value     |
-| ------------ | ------------ |
-| train data   | 9            |
-| train size   | 6_370_521    |
-| val size     | 1_124_210    |
-| train mean   | 0.5168       |
-| train min    | 0            |
-| train max    | 1            |
-| t-near 0.5   | 0.2%         |
-| val mean     | 0.5170       |
-| val min      | 0            |
-| val max      | 1            |
-| v-near 0.5   | 0.2%         |
-| e1 t-loss    | 0.35914      |
-| e1 v-loss    | 0.34719      |
-| e197 t-loss  | 0.33796      |
-| e197 v-loss  | 0.33629      |
-| val gap      | 0.00167      |
-
-### Tournaments
-
-===========================================
-  Tournament Results gen9 vs. gen8
-===========================================
-  Model A : training_net (gen9)
-  Model B : training_net (gen8)
-  Modus   : (from game)
-  Total games  : 2000
-  Decisive     : 2000
-  Draws        : 0
-  Discarded    : 0
-  Avg turns    : 98.0
-  Model A wins : 1062 (53.1%)
-  Model B wins : 938 (46.9%)
-  A win rate   : 53.10%
-  95% CI       : [50.91%, 55.28%]
-  Significance : p<0.01  (z=2.77) - significant
-  Win rate last 10 checkpoints: 53.17% ± 0.04%
-  Verdict: INCONCLUSIVE.
-===========================================
