+++ src/GammonX/GammonX.Mars.Server/NeuralNets/Backgammon/README.md
@@ -1,0 +1,32 @@
+### Trainings Data
+|    Metric    |    Value     |
+| ------------ | ------------ |
+| prior gen    | linear model |
+| sample size  | 80_000       |
+| bootstrap    | 1            |
+| completed    | 80_000       |
+| discarded    | 0            |
+| samples      | 4_807_019    |
+| avg turns    | 64.2         |
+| avg pred var | -            |
+| duration     | 00:00:56:24  |
+
+### Training Model
+|    Metric    |    Value     |
+| ------------ | ------------ |
+| train data   | 0            |
+| train size   | 4_085_966    |
+| val size     | 721_053      |
+| train mean   | 0.5187       |
+| train min    | 0            |
+| train max    | 1            |
+| t-near 0.5   | 0.=%         |
+| val mean     | 0.5179       |
+| val min      | 0            |
+| val max      | 1            |
+| v-near 0.5   | 0.0%         |
+| e1 t-loss    | 0.21596      |
+| e1 v-loss    | 0.20141      |
+| e196 t-loss  | 0.19697      |
+| e196 v-loss  | 0.19523      |
+| val gap      | 0.00174      |
